Changed priority type

+++ DataStructures/Interfaces/IPriorityQueueItem.cs
@@ -2,6 +2,6 @@
 {
 	public interface IPriorityQueueItem
 	{
-		int Priority { get; set; }
+		double Priority { get; set; }
 	}
 }
